+++ src/__tests__/utils/index.test.ts
@@ -1,19 +1,5 @@
-import {
-  getDeepValue,
-  getUnique,
-  getUniqueBy,
-  serialize
-} from '../../utils/getUniqueBy'
-import { belongsTo } from '../../utils/functions'
-import { isEqual } from '../../utils/isEqual'
+import { getUnique, getUniqueBy, isOneOf, isEqual } from '../../../dist/utils'
 
 import { commonUtilTests } from '.'
 
-commonUtilTests({
-  belongsTo,
-  getDeepValue,
-  getUnique,
-  getUniqueBy,
-  serialize,
-  isEqual
-})
+commonUtilTests({ isOneOf, getUnique, getUniqueBy, isEqual })