+++ src/__tests__/utils/index.ts
@@ -1,34 +1,30 @@
-interface CommonUtilsProps {
-  [key: string]: Function
-}
+type CommonUtilsProps = { [key: string]: Function }
 
 export const commonUtilTests = ({
-  belongsTo,
-  getDeepValue,
+  isOneOf,
   getUnique,
   getUniqueBy,
-  isEqual,
-  serialize
+  isEqual
 }: CommonUtilsProps) => {
   describe('belongsTo', () => {
     it('should return true if value passed is in array supplied else false', () => {
       const values = [1, 'hey', null, undefined, false]
 
       // truthy tests
-      expect(belongsTo(1, values)).toBe(true)
-      expect(belongsTo('hey', values)).toBe(true)
-      expect(belongsTo(null, values)).toBe(true)
-      expect(belongsTo(undefined, values)).toBe(true)
-      expect(belongsTo(false, values)).toBe(true)
+      expect(isOneOf(1, values)).toBe(true)
+      expect(isOneOf('hey', values)).toBe(true)
+      expect(isOneOf(null, values)).toBe(true)
+      expect(isOneOf(undefined, values)).toBe(true)
+      expect(isOneOf(false, values)).toBe(true)
 
       // falsy tests
-      expect(belongsTo('1', values)).toBe(false)
-      expect(belongsTo('Hey', values)).toBe(false)
-      expect(belongsTo('null', values)).toBe(false)
-      expect(belongsTo('undefined', values)).toBe(false)
-      expect(belongsTo('false', values)).toBe(false)
-      expect(belongsTo(2, values)).toBe(false)
-      expect(belongsTo(true, values)).toBe(false)
+      expect(isOneOf('1', values)).toBe(false)
+      expect(isOneOf('Hey', values)).toBe(false)
+      expect(isOneOf('null', values)).toBe(false)
+      expect(isOneOf('undefined', values)).toBe(false)
+      expect(isOneOf('false', values)).toBe(false)
+      expect(isOneOf(2, values)).toBe(false)
+      expect(isOneOf(true, values)).toBe(false)
     })
   })
 
@@ -42,7 +38,13 @@ export const commonUtilTests = ({
       expect(isEqual([1, 'true', [], null], [1, 'true', [], null])).toEqual(
         true
       )
-      expect(isEqual({ name: 'James' }, { name: 'James' })).toEqual(true)
+      expect(isEqual({ a: 'James' }, { a: 'James' })).toEqual(true)
+      expect(isEqual({ a: '' }, { a: '' })).toEqual(true)
+      expect(isEqual({ a: '', b: '' }, { a: '', b: '' })).toEqual(true)
+      expect(isEqual({ a: '', b: '' }, { b: '', a: '' })).toEqual(true)
+      expect(isEqual({ a: '', b: { c: '' } }, { b: { c: '' }, a: '' })).toEqual(
+        true
+      )
 
       // falsy
       expect(isEqual(1, '1')).toEqual(false)
@@ -51,56 +53,89 @@ export const commonUtilTests = ({
       expect(isEqual([1, 'true', [], null], [1, 'true', null, []])).toEqual(
         false
       )
-      expect(isEqual({ name: 'James' }, { name: 'JameS' })).toEqual(false)
-      expect(isEqual({ name: 'James' }, { name: 'James', age: 17 })).toEqual(
-        false
+      expect(isEqual({ a: 'James' }, { a: 'JameS' })).toEqual(false)
+      expect(isEqual({ a: 'James' }, { a: 'James', b: 17 })).toEqual(false)
+    })
+
+    it('should respect the level of nesting(depth)', () => {
+      // depth == undefined (defaults to 1)
+
+      for (const depth of [undefined, 1]) {
+        expect(
+          isEqual({ a: '', b: { c: '' } }, { b: { c: '' }, a: '' }, depth)
+        ).toEqual(true)
+
+        expect(
+          isEqual({ a: '', b: [1, 2] }, { b: [1, 2], a: '' }, depth)
+        ).toEqual(true)
+
+        expect(
+          isEqual(
+            { a: '', b: { c: '', d: [1, 2] } },
+            { b: { d: [1, 2], c: '' }, a: '' },
+            depth
+          )
+        ).toEqual(true)
+
+        expect(
+          isEqual(
+            { a: '', b: { d: '', c: '' } },
+            { b: { c: '', d: '' }, a: '' },
+            depth
+          )
+        ).toEqual(true)
+      }
+
+      // depth == 0
+      expect(
+        isEqual({ a: '', b: { c: '' } }, { b: { c: '' }, a: '' }, 0)
+      ).toEqual(true)
+
+      expect(isEqual({ a: '', b: [1, 2] }, { b: [1, 2], a: '' }, 0)).toEqual(
+        true
       )
-    })
-  })
 
-  describe('getDeepValue', () => {
-    let person: any
+      expect(
+        isEqual(
+          { a: '', b: { c: '', d: [1, 2] } },
+          { b: { d: [1, 2], c: '' }, a: '' },
+          0
+        )
+      ).toEqual(false)
 
-    beforeAll(() => {
-      person = {
-        name: 'James',
-        age: 20,
-        bio: {
-          joinDate: 'today',
-          facebook: { link: '/facebook/james', likes: 1700 }
-        }
+      expect(
+        isEqual(
+          { a: '', b: { d: '', c: '' } },
+          { b: { c: '', d: '' }, a: '' },
+          0
+        )
+      ).toEqual(false)
+
+      for (const depth of [2, 3, Infinity]) {
+        expect(
+          isEqual({ a: '', b: { c: '' } }, { b: { c: '' }, a: '' }, depth)
+        ).toEqual(true)
+
+        expect(
+          isEqual({ a: '', b: [1, 2] }, { b: [1, 2], a: '' }, depth)
+        ).toEqual(true)
+
+        expect(
+          isEqual(
+            { a: '', b: { c: '', d: [1, 2] } },
+            { b: { d: [1, 2], c: '' }, a: '' },
+            depth
+          )
+        ).toEqual(true)
+
+        expect(
+          isEqual(
+            { a: '', b: { d: '', c: '' } },
+            { b: { c: '', d: '' }, a: '' },
+            depth
+          )
+        ).toEqual(true)
       }
-    })
-
-    it('should give value with simple keys', () => {
-      const truthy: [string, any][] = [
-        ['name', 'James'],
-        ['age', 20]
-      ]
-
-      for (const [key, value] of truthy) {
-        expect(getDeepValue(person, key)).toBe(value)
-      }
-    })
-
-    it('should give value with nested keys', () => {
-      const truthy: [string, any][] = [
-        ['bio.joinDate', 'today'],
-        ['bio.facebook.link', '/facebook/james'],
-        ['bio.facebook.likes', 1700]
-      ]
-
-      for (const [key, value] of truthy) {
-        expect(getDeepValue(person, key)).toBe(value)
-      }
-    })
-
-    it('should give undefined if simple key is not set', () => {
-      expect(getDeepValue(person, 'dob')).toBe(undefined)
-    })
-
-    it('should give undefined if nested key is not set', () => {
-      expect(getDeepValue(person, 'address.streetName')).toBe(undefined)
     })
   })
 
@@ -142,36 +177,6 @@ export const commonUtilTests = ({
 
       expect(getUniqueBy(values, 'name').length).toBe(2)
       expect(getUniqueBy(values, 'age').length).toBe(1)
-    })
-  })
-
-  describe('serialize', () => {
-    it('should convert values to json strings', () => {
-      const truthy = [
-        ['1', '"1"'],
-        [1, '1'],
-        [{}, '{}'],
-        [[], '[]']
-      ]
-
-      for (const [key, value] of truthy) {
-        expect(serialize(key)).toBe(value)
-      }
-    })
-
-    it('should convert json values their original types', () => {
-      const truthy = [
-        ['"1"', '1'],
-        ['1', 1]
-      ]
-
-      for (const [key, value] of truthy) {
-        expect(serialize(key, true)).toBe(value)
-      }
-    })
-
-    it('should not serialize nor deserialise undefined', () => {
-      expect(serialize(undefined)).toBe(serialize(undefined, true))
     })
   })
 }